// ui/src/generated/FileSearchResult.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DirectoryEntry } from "./DirectoryEntry";

export type FileSearchResult = { entries: Array<DirectoryEntry>, scanned_entries: number, scanned_directories: number, truncated: boolean, notices: Array<string>, };